requirements.txt
numpy==1.26.4
torch==2.2.2
transformers==4.45.2
peft==0.17.1
accelerate==0.34.2
datasets==3.0.1
tqdm==4.66.5
llama-cpp-python==0.2.90
plotly==5.24.1
tokenizers==0.19.1
